Shell commands working
- reorg code
- fix ui bugs
